moved GuiApp and associated files to view, ab#797
