+++ pom.xml
@@ -106,7 +106,7 @@
                         <!-- Default configuration for running with: mvn clean javafx:run -->
                         <id>default-cli</id>
                         <configuration>
-                            <mainClass>controller.demo1/controller.GuiApp</mainClass>
+                            <mainClass>controller.demo1/view.GuiApp</mainClass>
                         </configuration>
                     </execution>
                 </executions>
