+++ src/main/java/view/GuiApp.java
@@ -1,4 +1,4 @@
-package controller;
+package view;
 
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
