+++ src/main/resources/view/intro.png
